requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.1.0
psycopg2-binary==2.9.10
djangorestframework==3.15.0
